refactor(azure-firewall): drop redundant 'azure' dir prefix (docs/STRUCTURE.md §2.2)
