+++ providers/azure/azure.go
@@ -13,7 +13,6 @@ import (
 	"github.com/stackshy/cloudemu/v2/providers/azure/ai"
 	"github.com/stackshy/cloudemu/v2/providers/azure/aks"
 	"github.com/stackshy/cloudemu/v2/providers/azure/applicationgateway"
-	"github.com/stackshy/cloudemu/v2/providers/azure/azurefirewall"
 	"github.com/stackshy/cloudemu/v2/providers/azure/blobstorage"
 	"github.com/stackshy/cloudemu/v2/providers/azure/cache"
 	"github.com/stackshy/cloudemu/v2/providers/azure/containerapps"
@@ -23,6 +22,7 @@ import (
 	"github.com/stackshy/cloudemu/v2/providers/azure/databricks"
 	"github.com/stackshy/cloudemu/v2/providers/azure/dns"
 	"github.com/stackshy/cloudemu/v2/providers/azure/eventgrid"
+	"github.com/stackshy/cloudemu/v2/providers/azure/firewall"
 	"github.com/stackshy/cloudemu/v2/providers/azure/functions"
 	"github.com/stackshy/cloudemu/v2/providers/azure/iam"
 	"github.com/stackshy/cloudemu/v2/providers/azure/keyvault"
@@ -136,7 +136,7 @@ type Provider struct {
 	DNS              *dns.Mock
 	LB               *loadbalancer.Mock
 	AppGateway       *applicationgateway.Mock
-	Firewall         *azurefirewall.Mock
+	Firewall         *firewall.Mock
 	ServiceBus       *servicebus.Mock
 	// QueueStorage backs the Azure Queue Storage data-plane handler. It reuses
 	// the messagequeue provider, but is a distinct instance from ServiceBus so
@@ -199,7 +199,7 @@ func New(opts ...config.Option) *Provider {
 		DNS:                dns.New(o),
 		LB:                 loadbalancer.New(o),
 		AppGateway:         applicationgateway.New(o),
-		Firewall:           azurefirewall.New(o),
+		Firewall:           firewall.New(o),
 		ServiceBus:         servicebus.New(o),
 		QueueStorage:       servicebus.New(o),
 		TableStorage:       tablestorage.New(o),
