+++ server/azure/azure.go
@@ -18,7 +18,6 @@ import (
 	aksserver "github.com/stackshy/cloudemu/v2/server/azure/aks"
 	appinsightssrv "github.com/stackshy/cloudemu/v2/server/azure/appinsights"
 	appgatewaysrv "github.com/stackshy/cloudemu/v2/server/azure/applicationgateway"
-	azurefirewallsrv "github.com/stackshy/cloudemu/v2/server/azure/azurefirewall"
 	"github.com/stackshy/cloudemu/v2/server/azure/blobstorage"
 	cachesrv "github.com/stackshy/cloudemu/v2/server/azure/cache"
 	containerappssrv "github.com/stackshy/cloudemu/v2/server/azure/containerapps"
@@ -46,6 +45,7 @@ import (
 	dnssrv "github.com/stackshy/cloudemu/v2/server/azure/dns"
 	eventgridsrv "github.com/stackshy/cloudemu/v2/server/azure/eventgrid"
 	eventhubsrv "github.com/stackshy/cloudemu/v2/server/azure/eventhub"
+	azurefirewallsrv "github.com/stackshy/cloudemu/v2/server/azure/firewall"
 	"github.com/stackshy/cloudemu/v2/server/azure/functions"
 	"github.com/stackshy/cloudemu/v2/server/azure/iam"
 	"github.com/stackshy/cloudemu/v2/server/azure/images"
