+++ server/azure/firewall/policy_ops.go
@@ -1,4 +1,4 @@
-package azurefirewall
+package firewall
 
 import (
 	"net/http"
