+++ server/azure/firewall/sdk_roundtrip_test.go
@@ -1,4 +1,4 @@
-package azurefirewall_test
+package firewall_test
 
 import (
 	"context"
